### 2018-03-26 #636 @terrytangyuan
Sure. By doing a random search, it looks the following reviewers are related to "visualizations" and "R":
@kellieotto @zfurnas @jkarreth @corybrunson @chlalanne
