### 2018-03-28 #636 @arfon
@kellieotto @jkarreth @corybrunson @chlalanne - would you be willing to review this submission for JOSS?
