### 2018-03-28 #636 @corybrunson
I certainly would! Though i would not be able to begin until the end of next week.
