### 2018-03-29 #636 @chlalanne
Sure I would! I should probably be able to review this article within 5-8 business days.

—
Christophe Lalanne

> Le 28 mars 2018 à 20:50, Arfon Smith <notifications@github.com> a écrit :
>
> @kellieotto @jkarreth @corybrunson @chlalanne - would you be willing to review this submission for JOSS?
>
