### 2018-03-29 #657 @whedon
[REVIEW]: autoplotly: An R package for automatic generation of interactive visualizations for statistical results
**Submitting author:** @terrytangyuan (<a href="http://orcid.org/0000-0001-5243-233X">Yuan Tang</a>)
**Repository:** <a href="https://github.com/terrytangyuan/autoplotly" target ="_blank">https://github.com/terrytangyuan/autoplotly</a>
**Version:** v0.1.1
