### 2018-03-29 #657 @terrytangyuan
Thanks @arfon. Do you know why the references do not show up in the compiled PDF?
