### 2018-03-29 #657 @arfon
@terrytangyuan - I think this is because they're not being cited directly. (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html))
