### 2018-03-29 f2cc66f
Directly cite bib entries
