+++ paper/paper.bib
@@ -3,20 +3,22 @@
     author = {Masaaki Horikoshi and Yuan Tang},
     year = {2015},
     note = {R package version 0.1.0},
-    url = {http://CRAN.R-project.org/package=ggfortify},
-  }
+    url = {http://CRAN.R-project.org/package=ggfortify}
+}
   
 @Manual{plotly,
     title = {plotly: Create Interactive Web Graphics via 'plotly.js'},
-    author = {Carson Sievert and Chris Parmer and Toby Hocking and Scott Chamberlain and Karthik Ram and Marianne Corvellec and Pedro Despouy},
-  }
+    year = {2018},
+    author = {Carson Sievert and Chris Parmer and Toby Hocking and Scott Chamberlain and Karthik Ram and Marianne Corvellec and Pedro Despouy}
+}
   
 @Manual{autoplotly,
     title = {autoplotly: Automatic Generation of Interactive Visualizations for Popular Statistical Results},
-    author = {Yuan Tang},
+    author = {Tang, Yuan},
     note = {R package version 0.1.0},
-    url = {https://github.com/terrytangyuan/autoplotly},
-  }
+    year = {2018},
+    url = {https://github.com/terrytangyuan/autoplotly}
+}
 
 @Article{rjggfortify,
   author = {Yuan Tang and Masaaki Horikoshi and Wenxuan Li},
@@ -25,7 +27,7 @@
   year = {2016},
   issue = {2},
   volume = {8},
-  url = {https://journal.r-project.org/},
+  url = {https://journal.r-project.org/}
 }
 
 @book{wickham2009ggplot2,
