### 2018-04-13 #657 @corybrunson
@terrytangyuan i have a question that i don't think is appropriate as an issue on the submitted repo. It may just reflect my lack of knowledge about software licenses.

**autoplotly** is licensed under Apache 2.0,  which is [described as incompatible with GPLv3 (and v2) in this direction](https://www.apache.org/licenses/GPL-compatibility.html). Since **ggplot2** and **ggfortify** are licensed under GPLv2, this seems to present a problem. Can you address that?
