### 2018-04-13 terrytangyuan/autoplotly#3 @corybrunson
package vignette from blog post?
The [blog post](https://terrytangyuan.github.io/2018/02/12/autoplotly-intro/) linked in the README provides a very helpful intro to the package and its intended usage. I suspect, especially since the package is on CRAN, that many R users would seek guidance from a vignette and miss the links to this resource (`DESCRIPTION` -> GitHub repo -> blog post). Could the blog post be repurposed into a package vignette?

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/657)._
