### 2018-04-13 terrytangyuan/autoplotly#4 @corybrunson
community guidelines
The `DESCRIPTION` contains a link to the issues page on the GitHub repo, but i did not see any other guidance on contributing to the package or seeking support.

The package consists of one core function whose job is to pass objects to `autoplot()` for which **ggfortify** provides methods. So users interested in contributing methods should presumably be pointed to [the **ggfortify** repo](https://github.com/sinhrks/ggfortify). (Presumably only experienced R users would want to contribute directly to the **autoplotly** source code.) But that repo also [has no contributing guidelines](https://github.com/sinhrks/ggfortify/community). If contributing to **ggfortify** would be as straightforward as it seems to me (a `fortify_*.R` file containing methods for `fortify()` and `autoplot()`), then the guide could be short; but there might be other important considerations i'm missing!

_Part of [this JOSS review](https://github.com/openjournals/joss-reviews/issues/657)._
