### 2018-04-13 terrytangyuan/autoplotly#4 @terrytangyuan
Thanks! I'll add those to the two repos.
