### 2018-04-13 terrytangyuan/autoplotly#3 @terrytangyuan
Thanks for pointing out! One problem with package vignette is that we cannot render interactive plots on CRAN. I think I'll just add the article as vignette with `eval = FALSE` then. Let me know what you think.
