### 2018-04-13 #657 @terrytangyuan
@corybrunson Thanks for taking the time to review! I just responded your issues. I've listed out the TODOs below:

- [ Move blogpost to package vignette (with `eval = FALSE` in .Rmd file)]
- [ Add community guidelines to ggfortify and autoplotly repos]
- [ Change autoplotly license to GPLv2]
- [ Add DOI to ggplot2 book citation]

I'll work on them soon.
