### 2018-04-13 #657 @terrytangyuan
@corybrunson Note that I am not adding a LICENSE file to the repo though since they are usually not required for R packages. This is a Github rendering issue and I believe they are working on the fix. I will fix the license in DESCRIPTION file only.
