### 2018-04-13 terrytangyuan/autoplotly#3 @corybrunson
Oh, that's a point i hadn't considered. Your suggestion sounds right—i'll take a look when it's ready (or on an experimental branch). Thanks!
