### 2018-04-13 124c4e3
Add community guidelines
