+++ CONTRIBUTING.md
@@ -1,0 +1,3 @@
+# Contributing guidelines
+
+Currently, `autoplotly` package is highly dependent on [ggfortify package](https://github.com/sinhrks/ggfortify) so most of the underlying implementation happens there. Please take a look ggfortify's contributing guideline [here](https://github.com/sinhrks/ggfortify/blob/master/CONTRIBUTING.md). This repository should mainly focus on issue tracking related to autoplotly or bug fixes regarding the current implementation to support extensibility using additional `ggplot2` or `plotly` syntax.
