+++ ISSUE_TEMPLATE.md
@@ -1,0 +1,25 @@
+Please go to Stack Overflow for help and support with the following tags:
+
+https://stackoverflow.com/questions/tagged/ggfortify
+https://stackoverflow.com/questions/tagged/autoplotly
+
+If you open a GitHub issue, here is our policy:
+
+1. It must be a bug, a feature request, or a significant problem with documentation (for small docs fixes please send a PR instead).
+2. The form below must be filled out.
+
+**Here's why we have that policy**: autoplotly developers respond to issues. We want to focus on work that benefits the whole community, e.g., fixing bugs and adding features. Support only helps individuals. GitHub also notifies thousands of people when issues are filed. We want them to see you communicating an interesting problem, rather than being redirected to Stack Overflow.
+
+------------------------
+
+### System information
+- **OS Platform and Distribution (e.g., Linux Ubuntu 16.04)**:
+- **ggfortify and autoplotly installed from (e.g. CRAN or Github)**:
+- **ggfortify and autoplotly version**:
+- **Exact command to reproduce**:
+
+### Describe the problem
+Describe the problem clearly here. Be sure to convey here why it's a bug in ggfortify/autoplotly or a feature request.
+
+### Source code / logs / plots
+Include any logs, source code, and plots that would be helpful to diagnose the problem. If including tracebacks, please include the full traceback. Large logs and files should be attached. Try to provide a reproducible test case that is the bare minimum necessary to generate the problem.
