+++ DESCRIPTION
@@ -10,7 +10,7 @@ Description: Functionalities to automatically generate interactive visualization
              clustering and survival analysis, with 'plotly.js' <https://plot.ly/>  and
              'ggplot2' style. The generated visualizations can also be easily extended
              using 'ggplot2' and 'plotly' syntax while staying interactive.
-License: Apache License 2.0
+License: GPL-2
 URL: https://github.com/terrytangyuan/autoplotly
 BugReports: https://github.com/terrytangyuan/autoplotly/issues
 Encoding: UTF-8
@@ -24,6 +24,7 @@ Imports:
     ggfortify
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 6.0.1
+VignetteBuilder: knitr
 Suggests:
     testthat,
     knitr
