### 2018-04-13 d86f292
Minor edits
