+++ vignettes/intro.Rmd
@@ -1,11 +1,11 @@
 <!--
 %\VignetteEngine{knitr::knitr}
-%\VignetteIndexEntry{Introduction to ggfortify package}
+%\VignetteIndexEntry{Introduction to autoplotly package}
 -->
 
 ```{r global_options, include=FALSE}
 library(knitr)
-opts_chunk$set(fig.width=6, fig.height=4, fig.path='figures/basics-', warning=FALSE, eval = FALSE)
+opts_chunk$set(warning=FALSE, eval = FALSE)
 ```
 
 ## Background
