+++ CONTRIBUTING.md
@@ -1,3 +1,23 @@
 # Contributing guidelines
 
-Currently, `autoplotly` package is highly dependent on [ggfortify package](https://github.com/sinhrks/ggfortify) so most of the underlying implementation happens there. Please take a look ggfortify's contributing guideline [here](https://github.com/sinhrks/ggfortify/blob/master/CONTRIBUTING.md). This repository should mainly focus on issue tracking related to autoplotly or bug fixes regarding the current implementation to support extensibility using additional `ggplot2` or `plotly` syntax.
+This package is still in an early development stage. Some details in the
+visualizations may not look as perfect yet. We welcome suggestions and
+contributions from others. With this package, researchers will hopefully spend
+less time focusing on learning `ggplot2` syntax or interactive visualization
+packages like `plotly`. Instead they can spend more time on their work and
+research. The source code of the package is located on
+[Github](https://github.com/terrytangyuan/autoplotly) where users can test out
+development versions of the package and provide feature requests, feedback and
+bug reports. We encourage you to submit your issues and pull requests to help us
+make this package better for the R community. In the future, we'll try to
+support a much wider range of objects through a larger coverage in `ggfortify`.
+If you find any new use cases, please don't hesitate to submit an issue on
+Github and join us in the development!
+
+Currently, this package is highly dependent on [ggfortify
+package](https://github.com/sinhrks/ggfortify) so most of the underlying
+implementation happens there. Please take a look ggfortify's contributing
+guideline located [here](https://github.com/sinhrks/ggfortify/blob/master/CONTRIBUTING.md). This
+repository should mainly focus on tracking issues related to autoplotly or bug
+fixes regarding the current implementation to support extensibility using
+additional `ggplot2` or `plotly` syntax.
