### 2018-04-13 f2e7c8d
Add doi to ggplot2 reference
