+++ paper/paper.bib
@@ -34,5 +34,6 @@
   title={ggplot2: Elegant Graphics for Data Analysis},
   author={Wickham, Hadley},
   year={2009},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media},
+  doi={10.1007/978-0-387-98141-3}
 }
