### 2018-04-13 #657 @corybrunson
@terrytangyuan Great! Everything looks and works good for me. I have one final suggestion for the README: Now that a vignette is included, the development version might better be installed via:

    devtools::install_github("terrytangyuan/autoplotly", build_vignettes = TRUE)
