### 2018-04-19 #657 @chlalanne
Sorry for the delay. I somewhat messed up with GH notifications. The package looks great and is working out of the box for me. I reviewed the vignette and the test suite, and they look ok to me.

@terrytangyuan I only have a couple a minor comments:

1. You cited SO as a possible placeholder for reporting issues or seeking help (`ISSUE_TEMPLATE.md`); maybe you should also indicate that users need to provide a minimal reproducible example, per SO policy, to prevent questions being closed quickly.
2. You do not mention `plotly::ggplotly()`, which provides similar facilities but without the `ggfortify` backend. I would have thought that it would perhaps fit in the last paragraph of you `paper.md` file.

@arfon I am done with the review and I am happy with the project as it stands.
