+++ paper/paper.bib
@@ -20,6 +20,13 @@
     url = {https://github.com/terrytangyuan/autoplotly}
 }
 
+@article{autoplotlyarxiv,
+  title={Autoplotly - Automatic Generation of Interactive Visualizations for Popular Statistical Results},
+  author={Tang, Yuan},
+  journal={arXiv preprint arXiv:1803.08424},
+  year={2018}
+}
+
 @Article{rjggfortify,
   author = {Yuan Tang and Masaaki Horikoshi and Wenxuan Li},
   title = {ggfortify: Unified Interface to Visualize Statistical Result of Popular R Packages},
