### 2018-04-19 #657 @arfon
Great! @terrytangyuan - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
