+++ paper/paper.bib
@@ -20,7 +20,7 @@
     url = {https://github.com/terrytangyuan/autoplotly}
 }
 
-@article{autoplotlyarxiv,
+@Article{autoplotlyarxiv,
   title={Autoplotly - Automatic Generation of Interactive Visualizations for Popular Statistical Results},
   author={Tang, Yuan},
   journal={arXiv preprint arXiv:1803.08424},
@@ -28,12 +28,13 @@
 }
 
 @Article{rjggfortify,
-  author = {Yuan Tang and Masaaki Horikoshi and Wenxuan Li},
+  author = {Tang, Yuan and Horikoshi, Masaaki and Li, Wenxuan},
   title = {ggfortify: Unified Interface to Visualize Statistical Result of Popular R Packages},
   journal = {The R Journal},
   year = {2016},
   issue = {2},
   volume = {8},
+  pages={478--489},
   url = {https://journal.r-project.org/}
 }
 
