### 2018-04-19 #657 @terrytangyuan
@arfon Please hold off on the publishing. I've resolving some minor issues with the Rmarkdown generation. Will ping you when ready. BTW, just curious - if I update paper.md going forward, would it automatically update the paper?
