+++ paper/paper.bib
@@ -1,8 +1,8 @@
 @Manual{ggfortify,
     title = {ggfortify: Data Visualization Tools for Statistical Analysis Results},
-    author = {Masaaki Horikoshi and Yuan Tang},
-    year = {2015},
-    note = {R package version 0.1.0},
+    author = {Horikoshi, Masaaki and Tang, Yuan},
+    year = {2018},
+    note = {R package version 0.4.4},
     url = {http://CRAN.R-project.org/package=ggfortify}
 }
   
@@ -12,7 +12,7 @@
     author = {Carson Sievert and Chris Parmer and Toby Hocking and Scott Chamberlain and Karthik Ram and Marianne Corvellec and Pedro Despouy}
 }
   
-@Manual{autoplotly,
+@Manual{autoplotlypkg,
     title = {autoplotly: Automatic Generation of Interactive Visualizations for Popular Statistical Results},
     author = {Tang, Yuan},
     note = {R package version 0.1.0},
