### 2018-04-19 #657 @terrytangyuan
@arfon All issues fixed now. See DOI in my comments above (you can also find it in project README). Thanks for everyone’s time!
