### 2018-04-20 #657 @arfon
> BTW, just curious - if I update paper.md going forward, would it automatically update the paper?

No, it wouldn't update on the final JOSS site but @whedon can always compile the latest version for you here.
